Re-introduce the "old-style" dog-ear "zoom"-icon via another appearance-option.
The initial version of the theme – which used CSS-background-images instead of webfonts for all icons – used a small "dog-ear"-like icon in the top right corner of images with full-size-versions to offer a "zoom"-option to the user. With the introduction of webfonts-as-icons, this zoom-button got much more prominent, being more than twice the size and centered in the image to be zoomed.

With this commit, the much calmer "dog-ear"-variant is optionally available again. A little rough around the edges still, but we're getting there.